requirements.txt
chromadb==0.4.0
torch==2.3.0
torchvision==0.18.0
numpy>=2.0.0
Pillow>=10.0.0
tqdm>=4.65.0
git+https://github.com/openai/CLIP.git
